refactor: Split farm hooks
