+++ src/state/hooks.ts
@@ -7,14 +7,9 @@ import { minBy, orderBy } from 'lodash'
 import { useAppDispatch } from 'state'
 import { Team } from 'config/constants/types'
 import Nfts from 'config/constants/nfts'
-import { farmsConfig } from 'config/constants'
 import { simpleRpcProvider } from 'utils/providers'
-import { getBalanceAmount } from 'utils/formatBalance'
-import { BIG_ZERO } from 'utils/bigNumber'
 import useRefresh from 'hooks/useRefresh'
-import { filterFarmsByQuoteToken } from 'utils/farmsPriceHelpers'
 import {
-  fetchFarmsPublicDataAsync,
   fetchPoolsPublicDataAsync,
   fetchPoolsUserDataAsync,
   fetchCakeVaultPublicData,
@@ -24,12 +19,10 @@ import {
 } from './actions'
 import {
   State,
-  Farm,
   Pool,
   ProfileState,
   TeamsState,
   AchievementState,
-  FarmsState,
   NodeRound,
   ReduxNodeLedger,
   NodeLedger,
@@ -42,7 +35,6 @@ import { fetchWalletNfts } from './collectibles'
 import { parseBigNumberObj } from './predictions/helpers'
 import { transformPool } from './pools/helpers'
 import { fetchPoolsStakingLimitsAsync } from './pools'
-import { fetchFarmUserDataAsync, nonArchivedFarms } from './farms'
 import {
   fetchCurrentLotteryId,
   fetchCurrentLottery,
@@ -50,37 +42,7 @@ import {
   fetchPublicLotteries,
 } from './lottery'
 import { useProcessLotteryResponse } from './lottery/helpers'
-
-export const usePollFarmsData = (includeArchive = false) => {
-  const dispatch = useAppDispatch()
-  const { slowRefresh } = useRefresh()
-  const { account } = useWeb3React()
-
-  useEffect(() => {
-    const farmsToFetch = includeArchive ? farmsConfig : nonArchivedFarms
-    const pids = farmsToFetch.map((farmToFetch) => farmToFetch.pid)
-
-    dispatch(fetchFarmsPublicDataAsync(pids))
-
-    if (account) {
-      dispatch(fetchFarmUserDataAsync({ account, pids }))
-    }
-  }, [includeArchive, dispatch, slowRefresh, account])
-}
-
-/**
- * Fetches the "core" farm data used globally
- * 251 = CAKE-BNB LP
- * 252 = BUSD-BNB LP
- */
-export const usePollCoreFarmData = () => {
-  const dispatch = useAppDispatch()
-  const { fastRefresh } = useRefresh()
-
-  useEffect(() => {
-    dispatch(fetchFarmsPublicDataAsync([251, 252]))
-  }, [dispatch, fastRefresh])
-}
+import { useFarmFromPid } from './farms/hooks'
 
 export const usePollBlockNumber = () => {
   const dispatch = useAppDispatch()
@@ -93,71 +55,6 @@ export const usePollBlockNumber = () => {
 
     return () => clearInterval(interval)
   }, [dispatch])
-}
-
-// Farms
-
-export const useFarms = (): FarmsState => {
-  const farms = useSelector((state: State) => state.farms)
-  return farms
-}
-
-export const useFarmFromPid = (pid): Farm => {
-  const farm = useSelector((state: State) => state.farms.data.find((f) => f.pid === pid))
-  return farm
-}
-
-export const useFarmFromLpSymbol = (lpSymbol: string): Farm => {
-  const farm = useSelector((state: State) => state.farms.data.find((f) => f.lpSymbol === lpSymbol))
-  return farm
-}
-
-export const useFarmUser = (pid) => {
-  const farm = useFarmFromPid(pid)
-
-  return {
-    allowance: farm.userData ? new BigNumber(farm.userData.allowance) : BIG_ZERO,
-    tokenBalance: farm.userData ? new BigNumber(farm.userData.tokenBalance) : BIG_ZERO,
-    stakedBalance: farm.userData ? new BigNumber(farm.userData.stakedBalance) : BIG_ZERO,
-    earnings: farm.userData ? new BigNumber(farm.userData.earnings) : BIG_ZERO,
-  }
-}
-
-// Return a farm for a given token symbol. The farm is filtered based on attempting to return a farm with a quote token from an array of preferred quote tokens
-export const useFarmFromTokenSymbol = (tokenSymbol: string, preferredQuoteTokens?: string[]): Farm => {
-  const farms = useSelector((state: State) => state.farms.data.filter((farm) => farm.token.symbol === tokenSymbol))
-  const filteredFarm = filterFarmsByQuoteToken(farms, preferredQuoteTokens)
-  return filteredFarm
-}
-
-// Return the base token price for a farm, from a given pid
-export const useBusdPriceFromPid = (pid: number): BigNumber => {
-  const farm = useFarmFromPid(pid)
-  return farm && new BigNumber(farm.token.busdPrice)
-}
-
-export const useBusdPriceFromToken = (tokenSymbol: string): BigNumber => {
-  const tokenFarm = useFarmFromTokenSymbol(tokenSymbol)
-  const tokenPrice = useBusdPriceFromPid(tokenFarm?.pid)
-  return tokenPrice
-}
-
-export const useLpTokenPrice = (symbol: string) => {
-  const farm = useFarmFromLpSymbol(symbol)
-  const farmTokenPriceInUsd = useBusdPriceFromPid(farm.pid)
-  let lpTokenPrice = BIG_ZERO
-
-  if (farm.lpTotalSupply && farm.lpTotalInQuoteToken) {
-    // Total value of base token in LP
-    const valueOfBaseTokenInFarm = farmTokenPriceInUsd.times(farm.tokenAmountTotal)
-    // Double it to get overall value in LP
-    const overallValueOfAllTokensInFarm = valueOfBaseTokenInFarm.times(2)
-    // Divide total value of all tokens, by the number of LP tokens
-    const totalLpTokens = getBalanceAmount(new BigNumber(farm.lpTotalSupply))
-    lpTokenPrice = overallValueOfAllTokensInFarm.div(totalLpTokens)
-  }
-
-  return lpTokenPrice
 }
 
 // Pools
