+++ src/views/Farms/components/FarmTable/Actions/StakedAction.tsx
@@ -6,7 +6,7 @@ import { BigNumber } from 'bignumber.js'
 import UnlockButton from 'components/UnlockButton'
 import Balance from 'components/Balance'
 import { useWeb3React } from '@web3-react/core'
-import { useFarmUser, useLpTokenPrice } from 'state/hooks'
+import { useFarmUser, useLpTokenPrice } from 'state/farms/hooks'
 import { fetchFarmUserDataAsync } from 'state/farms'
 import { FarmWithStakedValue } from 'views/Farms/components/FarmCard/FarmCard'
 import { useTranslation } from 'contexts/Localization'
