+++ src/views/Home/components/EarnAPRCard.tsx
@@ -7,7 +7,8 @@ import { NavLink } from 'react-router-dom'
 import BigNumber from 'bignumber.js'
 import { useTranslation } from 'contexts/Localization'
 import { useAppDispatch } from 'state'
-import { useFarms, usePriceCakeBusd } from 'state/hooks'
+import { usePriceCakeBusd } from 'state/hooks'
+import { useFarms } from 'state/farms/hooks'
 import { fetchFarmsPublicDataAsync, nonArchivedFarms } from 'state/farms'
 import { getFarmApr } from 'utils/apr'
 import useIntersectionObserver from 'hooks/useIntersectionObserver'
