+++ src/views/Ifos/hooks/v1/useGetPublicIfoData.ts
@@ -2,7 +2,8 @@ import { useEffect, useState, useCallback } from 'react'
 import BigNumber from 'bignumber.js'
 import { BSC_BLOCK_TIME } from 'config'
 import { Ifo, IfoStatus, PoolIds } from 'config/constants/types'
-import { useBlock, useLpTokenPrice } from 'state/hooks'
+import { useBlock } from 'state/hooks'
+import { useLpTokenPrice } from 'state/farms/hooks'
 import { BIG_ZERO } from 'utils/bigNumber'
 import { multicallv2 } from 'utils/multicall'
 import ifoV1Abi from 'config/abi/ifoV1.json'
